--- a/halo2_proofs/Cargo.toml
+++ b/halo2_proofs/Cargo.toml
@@ -32,6 +32,10 @@ name = "hashtocurve"
 harness = false
 
 [[bench]]
+name = "msm"
+harness = false
+
+[[bench]]
 name = "plonk"
 harness = false